merge

+++ UniGames.Api/Controllers/ReviewController.cs
@@ -1,12 +1,12 @@
 using AutoMapper;
-
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using UniGames.Api.Data;
 using UniGames.Api.Models.Domain;
 using UniGames.Api.Models.DTOs;
 using UniGames.Api.Repositories;
-
+using BCryptNet = BCrypt.Net.BCrypt;
 
 namespace UniGames.Api.Controllers
 {
@@ -71,6 +71,7 @@ namespace UniGames.Api.Controllers
 
         // Uses the HttpPost method
         [HttpPost]
+        //[Authorize]
         // Public Method
         public IActionResult CreateReview([FromBody] CreateReviewDTO createReviewDTO)
         {
@@ -78,6 +79,13 @@ namespace UniGames.Api.Controllers
             {
                 // Map DTO to DM
                 var reviewDM = mapper.Map<Review>(createReviewDTO);
+
+                var reviewExists = reviewRepository.GetReviewByUser(reviewDM.UserID);
+                if (reviewExists.Count > 1)
+                {
+                    return Conflict();
+                }
+
                 // Execute the Create Review Method
                 var crreview = reviewRepository.CreateReview(reviewDM);
 
@@ -89,9 +97,55 @@ namespace UniGames.Api.Controllers
             else
             {
                 return StatusCode(422, ModelState);
+                
             }
             
 
+        }
+
+        [HttpPut]
+        [Route("/editreview/{id:int}")]
+        public IActionResult EditReview([FromRoute] int id, [FromBody] UpdateReviewDTO updateReviewDTO)
+        {
+            if (ModelState.IsValid)
+            {
+                var reviewDM = reviewRepository.GetReviewByID(id);
+                if (reviewDM == null)
+                {
+                    return NotFound();
+                }
+                if (reviewDM.UserID != updateReviewDTO.UserID)
+                {
+                    return Unauthorized("User is editing the wrong review");
+                }
+                /*if (reviewDM.GameID != updateReviewDTO.GameID)
+                {
+                    return BadRequest("User is attempting to update the game ID for this review");
+                }*/
+
+                this.mapper.Map(updateReviewDTO, reviewDM);
+                dbContext.SaveChanges();
+
+                var reviewDTO = mapper.Map<ReviewDTO>(reviewDM);
+                return Ok(reviewDTO);
+            }
+            else
+            {
+                return StatusCode(500, new { Message = "Illegal method of editing a review attempted"});
+            }   
+        }
+
+        [HttpGet]
+        [Route("/userreview/{id:int}")]
+        public IActionResult GetReviewByUserID([FromRoute] int id)
+        {
+            var reviewDM = reviewRepository.GetReviewByUser(id);
+            if (reviewDM == null)
+            {
+                return NotFound();
+            }
+            var reviewDTO = mapper.Map<List<ReviewDTO>>(reviewDM);
+            return Ok(reviewDTO);
         }
 
         [HttpGet]
@@ -106,8 +160,9 @@ namespace UniGames.Api.Controllers
         }
 
         [HttpDelete]
-        [Route ("/deletereview/{userID:int}/{id:int}")]
-        public IActionResult DeleteReview([FromRoute] int userID, int id) {
+        [Route ("/deletereview/{userID:int}/{id:int}/{password}")]
+        //[Authorize]
+        public IActionResult DeleteReview([FromRoute] int userID, int id, string password) {
             
             var reviewdm = reviewRepository.GetReviewByID(id);
             if (reviewdm == null)
@@ -118,6 +173,15 @@ namespace UniGames.Api.Controllers
             {
                 return BadRequest();
             }
+
+            bool matchingPass = BCryptNet.Verify(password, reviewdm.UserName.Userpassword);
+            // If the password for the username does not match the database records then
+            if (!matchingPass)
+            {
+                // Return that the user is unauthorised
+                return Unauthorized();
+            }
+
 
             var delreview = reviewRepository.DeleteReview(reviewdm);
             var reviewdto = mapper.Map<ReviewDTO>(delreview);
